refactor(cli): trace cmd overhaul
